微修正 preparation から　const に　rename
